requirements.txt
attrs==22.2.0
exceptiongroup==1.0.4
greenlet==2.0.1
iniconfig==1.1.1
json3==1.0
loguru==0.6.0
lxml==4.9.2
numpy==1.23.5
packaging==22.0
pandas==1.5.2
pluggy==1.0.0
protobuf==3.20.1
psycopg==3.1.5
psycopg2-binary==2.9.5
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
tomli==2.0.1
typing_extensions==4.4.0
